daily/auto_api_docu/drf/requirements.txt
Django==2.0.4
django-rest-framework==0.1.0
djangorestframework==3.10.3
Pillow==6.1.0
pytz==2019.2
